Mirror 1.1.x change.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/branches/branch-1.2@3797 7a7537e8-13f0-0310-91df-b6672ffda945

+++ pdftops/OutputDev.h
@@ -72,7 +72,7 @@ public:
   //----- coordinate conversion
 
   // Convert between device and user coordinates.
-  virtual void cvtDevToUser(int dx, int dy, double *ux, double *uy);
+  virtual void cvtDevToUser(double dx, double dy, double *ux, double *uy);
   virtual void cvtUserToDev(double ux, double uy, int *dx, int *dy);
 
   //----- link borders
